requirements.txt
# データ処理
pandas==2.1.4
numpy==1.24.3
numba==0.58.1

# テクニカル分析
ta-lib==0.4.28
pandas-ta==0.3.14b0

# データベース
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# 可視化
plotly==5.18.0
dash==2.14.2
matplotlib==3.8.2

# バックテスト
scipy==1.11.4
statsmodels==0.14.1

# ユーティリティ
pydantic==2.5.3
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
black==23.12.1